+++ pages/_app.tsx
@@ -1,28 +1,12 @@
 import type { AppProps } from 'next/app'
 import '../styles/globals.css'
-import {
-  ChainMetadata,
-  SorobanReactProvider,
-  chain,
-  getDefaultWallets,
-} from "../wallet";
-
-const chains: ChainMetadata[] = [chain.sandbox, chain.standalone, chain.futurenet];
-
-const { wallets } = getDefaultWallets({
-  appName: "Example Stellar App",
-  chains
-});
+import ProviderExample from '../components/ProviderExample';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
-    <SorobanReactProvider
-      chains={chains}
-      appName={"Example Stellar App"}
-      wallets={wallets}
-      >
+    <ProviderExample>
       <Component {...pageProps} />
-    </SorobanReactProvider>
+    </ProviderExample>
   );
 }
